+++ backend/pick-git/src/test/java/com/woowacourse/pickgit/integration/post/PostServiceIntegrationTest_likeUsers.java
@@ -9,6 +9,7 @@ import com.woowacourse.pickgit.common.factory.UserFactory;
 import com.woowacourse.pickgit.config.InfrastructureTestConfiguration;
 import com.woowacourse.pickgit.exception.post.PostNotFoundException;
 import com.woowacourse.pickgit.post.application.PostFeedService;
+import com.woowacourse.pickgit.post.application.PostService;
 import com.woowacourse.pickgit.post.application.dto.request.AuthUserForPostRequestDto;
 import com.woowacourse.pickgit.post.application.dto.response.LikeUsersResponseDto;
 import com.woowacourse.pickgit.post.domain.Post;
@@ -32,10 +33,10 @@ import org.springframework.transaction.annotation.Transactional;
 @SpringBootTest(webEnvironment = WebEnvironment.NONE)
 @DirtiesContext(classMode = ClassMode.BEFORE_EACH_TEST_METHOD)
 @ActiveProfiles("test")
-public class PostFeedServiceIntegrationTest_likeUsers {
+public class PostServiceIntegrationTest_likeUsers {
 
     @Autowired
-    private PostFeedService postFeedService;
+    private PostService postService;
 
     @Autowired
     private UserRepository userRepository;
@@ -65,7 +66,7 @@ public class PostFeedServiceIntegrationTest_likeUsers {
             createLikeUserResponseDtoForLoginUser(likeUsers);
 
         // when
-        List<LikeUsersResponseDto> actualResponse = postFeedService
+        List<LikeUsersResponseDto> actualResponse = postService
             .likeUsers(authUserForPostRequestDto, postId);
 
         // then
@@ -96,7 +97,7 @@ public class PostFeedServiceIntegrationTest_likeUsers {
             createLikeUserResponseForGuest(likeUsers);
 
         // when
-        List<LikeUsersResponseDto> actualResponse = postFeedService
+        List<LikeUsersResponseDto> actualResponse = postService
             .likeUsers(authUserForPostRequestDto, postId);
 
         // then
@@ -119,7 +120,7 @@ public class PostFeedServiceIntegrationTest_likeUsers {
         Long postId = savedPost.getId();
 
         // when
-        List<LikeUsersResponseDto> actualResponse = postFeedService
+        List<LikeUsersResponseDto> actualResponse = postService
             .likeUsers(authUserForPostRequestDto, postId);
 
         // then
@@ -137,7 +138,7 @@ public class PostFeedServiceIntegrationTest_likeUsers {
 
         // when then
         assertThatThrownBy(
-            () -> postFeedService.likeUsers(authUserForPostRequestDto, postId)
+            () -> postService.likeUsers(authUserForPostRequestDto, postId)
         ).isInstanceOf(PostNotFoundException.class)
             .hasFieldOrPropertyWithValue("errorCode", "P0002");
     }
